add docs and move MapStatus to shuffle package
