+++ core/src/test/scala/org/apache/spark/ShuffleSuite.scala
@@ -67,7 +67,7 @@ class ShuffleSuite extends FunSuite with Matchers with LocalSparkContext {
 
     // All blocks must have non-zero size
     (0 until NUM_BLOCKS).foreach { id =>
-      val statuses = SparkEnv.get.shuffleManager.getServerStatuses(shuffleId, id)
+      val statuses = SparkEnv.get.shuffleManager.mapOutputTracker.getServerStatuses(shuffleId, id)
       assert(statuses.forall(s => s._2 > 0))
     }
   }
@@ -107,7 +107,7 @@ class ShuffleSuite extends FunSuite with Matchers with LocalSparkContext {
     assert(c.count === 4)
 
     val blockSizes = (0 until NUM_BLOCKS).flatMap { id =>
-      val statuses = SparkEnv.get.shuffleManager.getServerStatuses(shuffleId, id)
+      val statuses = SparkEnv.get.shuffleManager.mapOutputTracker.getServerStatuses(shuffleId, id)
       statuses.map(x => x._2)
     }
     val nonEmptyBlocks = blockSizes.filter(x => x > 0)
@@ -132,7 +132,7 @@ class ShuffleSuite extends FunSuite with Matchers with LocalSparkContext {
     assert(c.count === 4)
 
     val blockSizes = (0 until NUM_BLOCKS).flatMap { id =>
-      val statuses = SparkEnv.get.shuffleManager.getServerStatuses(shuffleId, id)
+      val statuses = SparkEnv.get.shuffleManager.mapOutputTracker.getServerStatuses(shuffleId, id)
       statuses.map(x => x._2)
     }
     val nonEmptyBlocks = blockSizes.filter(x => x > 0)
