+++ core/src/test/scala/org/apache/spark/shuffle/MapOutputTrackerSuite.scala
@@ -24,8 +24,6 @@ import akka.testkit.TestActorRef
 import org.scalatest.FunSuite
 
 import org.apache.spark._
-import org.apache.spark.scheduler.MapStatus
-import org.apache.spark.shuffle.FetchFailedException
 import org.apache.spark.storage.BlockManagerId
 import org.apache.spark.util.AkkaUtils
